Cleanup namespace and file structure (WIP)
