+++ src/batterymonitor.cpp
@@ -119,10 +119,10 @@ void BatteryMonitor::Loop()
                         #if defined(BATTERY_LOW_VOLTAGE_DEEP_SLEEP) && BATTERY_LOW_VOLTAGE_DEEP_SLEEP
                             ESP.deepSleep(0);
                         #else
-                            setLedStatus(LED_STATUS_LOW_BATTERY);
+                            LEDManager::setLedStatus(LED_STATUS_LOW_BATTERY);
                         #endif
                     } else {
-                        unsetLedStatus(LED_STATUS_LOW_BATTERY);
+                        LEDManager::unsetLedStatus(LED_STATUS_LOW_BATTERY);
                     }
                 #endif
             }
