+++ src/batterymonitor.h
@@ -25,10 +25,10 @@
 
 #include <Arduino.h>
 #include "defines.h"
-#include "udpclient.h"
+#include "network/network.h"
 #include <i2cscan.h>
 #include <i2cdev.h>
-#include "ledstatus.h"
+#include "ledmgr.h"
 
 #if BATTERY_MONITOR == BAT_EXTERNAL
     #ifndef PIN_BATTERY_LEVEL
