+++ src/ledmgr.h
@@ -26,11 +26,42 @@
 #include <Arduino.h>
 #include "defines.h"
 
-namespace LEDMGR {
-    void On(uint8_t pin);
-    void Off(uint8_t pin);
-    void Blink(uint8_t pin, unsigned long time, uint8_t direction = LOW);
-    void Pattern(uint8_t pin, unsigned long timeon, unsigned long timeoff, int times, uint8_t direction = LOW);
+#define LED_STATUS_SERVER_CONNECTING 2
+#define LED_STATUS_WIFI_CONNECTING 4
+#define LED_STATUS_LOW_BATTERY 128
+#define LED_STATUS_IMU_ERROR 256
+
+#define DEFAULT_LENGTH 300
+#define DEFAULT_GAP 500
+#define DEFAULT_INTERVAL 3000
+
+#define STANDBUY_LENGTH DEFAULT_LENGTH
+#define IMU_ERROR_LENGTH DEFAULT_LENGTH
+#define IMU_ERROR_INTERVAL 1000
+#define IMU_ERROR_COUNT 5
+#define LOW_BATTERY_LENGTH DEFAULT_LENGTH
+#define LOW_BATTERY_INTERVAL 300
+#define LOW_BATTERY_COUNT 1
+#define WIFI_CONNECTING_LENGTH DEFAULT_LENGTH
+#define WIFI_CONNECTING_INTERVAL 3000
+#define WIFI_CONNECTING_COUNT 3
+#define SERVER_CONNECTING_LENGTH DEFAULT_LENGTH
+#define SERVER_CONNECTING_INTERVAL 3000
+#define SERVER_CONNECTING_COUNT 2
+
+namespace LEDManager {
+    
+    enum Stage {
+        OFF, ON, GAP, INTERVAL
+    };
+    void on(uint8_t pin);
+    void off(uint8_t pin);
+    void blink(uint8_t pin, unsigned long time, uint8_t direction = LOW);
+    void pattern(uint8_t pin, unsigned long timeon, unsigned long timeoff, int times, uint8_t direction = LOW);
+    void signalAssert();
+    void setLedStatus(uint32_t status);
+    void unsetLedStatus(uint32_t status);
+    void ledStatusUpdate();
 }
 
 #endif // SLIMEVR_LEDMGR_H_
